COMMERCE-4433 Move test commerce context to correct package
